docs: fix norm
