+++ src/data_structure/linked_list_function_1.c
@@ -13,6 +13,19 @@
 #include "linked_list.h"
 #include "libft.h"
 
+t_list	*init_list(void)
+{
+	t_list	*list;
+
+	list = malloc(sizeof(t_list));
+	if (!list)
+		ft_terminate("init_list, malloc error");
+	list->head = NULL;
+	list->tail = NULL;
+	list->len = 0;
+	return (list);
+}
+
 t_node	*make_node(void *content)
 {
 	t_node	*new_node;
@@ -26,79 +39,22 @@ t_node	*make_node(void *content)
 	return (new_node);
 }
 
-void	pop_middle(t_list *list, t_node *node)
+void	swap_node(t_list *list)
 {
+	t_node	*fixed_node;
 	t_node	*cursor;
+	t_node	*cursor_l;
 
-	cursor = node;
-	cursor->prev->next = cursor->next;
-	cursor->next->prev = cursor->prev;
-	list->len--;
-	free(cursor);
-}
-
-void	pop_front(t_list *list)
-{
-	t_node	*cursor;
-
-	if (list->len == 0)
-		return ;
-	cursor = list->head;
-	list->head = cursor->next;
-	if (list->head == NULL)
-		list->tail = NULL;
-	else
-		list->head->prev = NULL;
-	list->len--;
-	free(cursor);
-}
-
-void	pop_back(t_list *list)
-{
-	t_node	*cursor;
-
-	if (list->len == 0)
-		return ;
 	cursor = list->tail;
-	list->tail = cursor->prev;
-	if (list->tail == NULL)
-		list->head = NULL;
-	else
-		list->tail->next = NULL;
-	list->len--;
-	free(cursor);
-}
-
-void	push_front(t_list *list, t_node *node)
-{
-	node->next = list->head;
-	node->prev = NULL;
-	if (list->head == NULL)
-	{
-		list->head = node;
-		list->tail = node;
-	}
-	else
-	{
-		list->head->prev = node;
-		list->head = node;
-	}
-	list->len++;
-}
-
-void	push_back(t_list *list, t_node *node)
-{
-	node->next = NULL;
-	node->prev = list->tail;
-	if (list->head == NULL)
-	{
-		list->head = node;
-		list->tail = node;
-	}
-	else
-	{
-		list->tail->next = node;
-		list->tail = node;
-	}
-	list->len++;
+	cursor_l = cursor->prev;
+	fixed_node = cursor_l->prev;
+	cursor->prev = fixed_node;
+	if (fixed_node != NULL)
+		fixed_node->next = cursor;
+	cursor->next = cursor_l;
+	cursor_l->prev = cursor;
+	cursor_l->next = NULL;
+	if (cursor_l == list->head)
+		list->head = cursor;
+	list->tail = cursor_l;
 }
