+++ src/data_structure/linked_list_function_2.c
@@ -13,49 +13,79 @@
 #include "linked_list.h"
 #include "libft.h"
 
-// void	clear_list(t_list *list)
-// {
-// 	t_node	*temp;
-
-// 	while (list->head != NULL)
-// 	{
-// 		temp = list->head->next;
-// 		free_content();
-// 		free(list->head->content);
-// 		free(list->head);
-// 		list->head = temp;
-// 	}
-// }
-
-void	swap_node(t_list *list)
+void	pop_front(t_list *list)
 {
-	t_node	*fixed_node;
 	t_node	*cursor;
-	t_node	*cursor_l;
 
-	cursor = list->tail;
-	cursor_l = cursor->prev;
-	fixed_node = cursor_l->prev;
-	cursor->prev = fixed_node;
-	if (fixed_node != NULL)
-		fixed_node->next = cursor;
-	cursor->next = cursor_l;
-	cursor_l->prev = cursor;
-	cursor_l->next = NULL;
-	if (cursor_l == list->head)
-		list->head = cursor;
-	list->tail = cursor_l;
+	if (list->len == 0)
+		return ;
+	cursor = list->head;
+	list->head = cursor->next;
+	if (list->head == NULL)
+		list->tail = NULL;
+	else
+		list->head->prev = NULL;
+	list->len--;
+	free(cursor);
 }
 
-t_list	*init_list(void)
+void	pop_back(t_list *list)
 {
-	t_list	*list;
+	t_node	*cursor;
 
-	list = malloc(sizeof(t_list));
-	if (!list)
-		ft_terminate("init_list, malloc error");
-	list->head = NULL;
-	list->tail = NULL;
-	list->len = 0;
-	return (list);
+	if (list->len == 0)
+		return ;
+	cursor = list->tail;
+	list->tail = cursor->prev;
+	if (list->tail == NULL)
+		list->head = NULL;
+	else
+		list->tail->next = NULL;
+	list->len--;
+	free(cursor);
+}
+
+void	pop_middle(t_list *list, t_node *node)
+{
+	t_node	*cursor;
+
+	cursor = node;
+	cursor->prev->next = cursor->next;
+	cursor->next->prev = cursor->prev;
+	list->len--;
+	free(cursor);
+}
+
+void	push_front(t_list *list, t_node *node)
+{
+	node->next = list->head;
+	node->prev = NULL;
+	if (list->head == NULL)
+	{
+		list->head = node;
+		list->tail = node;
+	}
+	else
+	{
+		list->head->prev = node;
+		list->head = node;
+	}
+	list->len++;
+}
+
+void	push_back(t_list *list, t_node *node)
+{
+	node->next = NULL;
+	node->prev = list->tail;
+	if (list->head == NULL)
+	{
+		list->head = node;
+		list->tail = node;
+	}
+	else
+	{
+		list->tail->next = node;
+		list->tail = node;
+	}
+	list->len++;
 }
